chore: refactor to simplify support for multiple API formats
