--- packages/rest/LICENSE
@@ -1,1 +1,0 @@
-../../LICENSE
