--- packages/rest/README.md
@@ -1,5 +1,0 @@
-# ZenStack REST Plugin Library
-
-This package provides a REST plugin for ZenStack.
-
-Visit [Homepage](https://zenstack.dev) for more details.
