--- packages/rest/package.json
@@ -1,52 +1,0 @@
-{
-    "name": "@zenstackhq/rest",
-    "version": "1.0.0-alpha.98",
-    "displayName": "ZenStack REST Adapters",
-    "description": "ZenStack REST adapters",
-    "homepage": "https://zenstack.dev",
-    "scripts": {
-        "clean": "rimraf dist",
-        "build": "pnpm lint && pnpm clean && tsc && copyfiles ./package.json ./README.md ./LICENSE dist",
-        "watch": "tsc --watch",
-        "lint": "eslint src --ext ts",
-        "test": "jest",
-        "prepublishOnly": "pnpm build",
-        "publish-dev": "pnpm publish --tag dev"
-    },
-    "publishConfig": {
-        "directory": "dist",
-        "linkDirectory": true
-    },
-    "keywords": [
-        "fastify"
-    ],
-    "author": "",
-    "license": "MIT",
-    "dependencies": {
-        "@zenstackhq/openapi": "workspace:*",
-        "@zenstackhq/runtime": "workspace:*",
-        "@zenstackhq/sdk": "workspace:*",
-        "@zenstackhq/server": "workspace:*",
-        "change-case": "^4.1.2",
-        "tiny-invariant": "^1.3.1",
-        "ts-japi": "^1.8.0",
-        "zod-validation-error": "^0.2.1"
-    },
-    "devDependencies": {
-        "@types/body-parser": "^1.19.2",
-        "@types/express": "^4.17.17",
-        "@types/jest": "^29.4.0",
-        "@types/supertest": "^2.0.12",
-        "@zenstackhq/testtools": "workspace:*",
-        "body-parser": "^1.20.2",
-        "copyfiles": "^2.4.1",
-        "express": "^4.18.2",
-        "fastify": "^4.14.1",
-        "fastify-plugin": "^4.5.0",
-        "jest": "^29.4.3",
-        "rimraf": "^3.0.2",
-        "supertest": "^6.3.3",
-        "ts-jest": "^29.0.5",
-        "typescript": "^4.9.4"
-    }
-}
