--- packages/rest/src/express/index.ts
@@ -1,1 +1,0 @@
-export { default as ZenStackMiddleware } from './middleware';
