--- packages/rest/src/express/middleware.ts
@@ -1,4 +1,0 @@
-import { handleRequest } from '../api';
-import { buildMiddlewareFactory } from '@zenstackhq/server/express';
-
-export default buildMiddlewareFactory(handleRequest);
