--- packages/rest/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "module": "CommonJS",
-        "lib": ["ESNext"],
-        "sourceMap": true,
-        "outDir": "dist",
-        "strict": true,
-        "noUnusedLocals": true,
-        "noImplicitReturns": true,
-        "moduleResolution": "node",
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "declaration": true,
-        "resolveJsonModule": true,
-        "strictPropertyInitialization": false,
-        "paths": {}
-    },
-    "include": ["src/**/*.ts"],
-    "exclude": ["dist", "node_modules"]
-}
