+++ packages/server/package.json
@@ -28,6 +28,7 @@
         "@zenstackhq/sdk": "workspace:*",
         "change-case": "^4.1.2",
         "tiny-invariant": "^1.3.1",
+        "ts-japi": "^1.8.0",
         "zod-validation-error": "^0.2.1"
     },
     "devDependencies": {
