+++ packages/server/src/express/index.ts
@@ -1,2 +1,1 @@
 export { default as ZenStackMiddleware } from './middleware';
-export { buildMiddlewareFactory } from './middleware';
