+++ packages/server/src/express/middleware.ts
@@ -2,12 +2,7 @@
 import { DbClientContract } from '@zenstackhq/runtime';
 import { getModelZodSchemas, ModelZodSchema } from '@zenstackhq/runtime/zod';
 import type { Handler, Request, Response } from 'express';
-import {
-    handleRequest as handleRequestOpenAPI,
-    LoggerConfig,
-    RequestContext,
-    Response as ResponseOpenAPI,
-} from '../openapi';
+import { HandleRequestFn, LoggerConfig } from '../api/utils';
 
 /**
  * Express middleware options
@@ -27,42 +22,42 @@ export interface MiddlewareOptions {
      * Zod schemas for validating request input. Pass `true` to load from standard location (need to enable `@core/zod` plugin in schema.zmodel).
      */
     zodSchemas?: ModelZodSchema | boolean;
-}
 
-type HandleRequestFn = (req: RequestContext) => Promise<ResponseOpenAPI>;
-type Factory = (options: MiddlewareOptions) => Handler;
+    /**
+     * API format to use from `@zenstackhq/server/api`
+     */
+    api: HandleRequestFn;
+}
 
 /**
  * Creates an Express middleware for handling CRUD requests.
  */
-export function buildMiddlewareFactory(handleRequest: HandleRequestFn): Factory {
-    return (options: MiddlewareOptions): Handler => {
-        let schemas: ModelZodSchema | undefined;
-        if (typeof options.zodSchemas === 'object') {
-            schemas = options.zodSchemas;
-        } else if (options.zodSchemas === true) {
-            schemas = getModelZodSchemas();
+const factory = (options: MiddlewareOptions): Handler => {
+    let schemas: ModelZodSchema | undefined;
+    if (typeof options.zodSchemas === 'object') {
+        schemas = options.zodSchemas;
+    } else if (options.zodSchemas === true) {
+        schemas = getModelZodSchemas();
+    }
+
+    return async (request, response) => {
+        const prisma = (await options.getPrisma(request, response)) as DbClientContract;
+        if (!prisma) {
+            throw new Error('unable to get prisma from request context');
         }
 
-        return async (request, response) => {
-            const prisma = (await options.getPrisma(request, response)) as DbClientContract;
-            if (!prisma) {
-                throw new Error('unable to get prisma from request context');
-            }
+        const r = await options.api({
+            method: request.method,
+            path: request.path,
+            query: request.query as Record<string, string | string[]>,
+            requestBody: request.body,
+            prisma,
+            logger: options.logger,
+            zodSchemas: schemas,
+        });
 
-            const r = await handleRequest({
-                method: request.method,
-                path: request.path,
-                query: request.query as Record<string, string | string[]>,
-                requestBody: request.body,
-                prisma,
-                logger: options.logger,
-                zodSchemas: schemas,
-            });
-
-            response.status(r.status).json(r.body);
-        };
+        response.status(r.status).json(r.body);
     };
-}
+};
 
-export default buildMiddlewareFactory(handleRequestOpenAPI);
+export default factory;
