+++ packages/server/src/fastify/plugin.ts
@@ -3,7 +3,7 @@ import { DbClientContract } from '@zenstackhq/runtime';
 import { getModelZodSchemas, ModelZodSchema } from '@zenstackhq/runtime/zod';
 import { FastifyPluginCallback, FastifyReply, FastifyRequest } from 'fastify';
 import fp from 'fastify-plugin';
-import { handleRequest, LoggerConfig } from '../openapi';
+import { HandleRequestFn, LoggerConfig } from '../api/utils';
 
 /**
  * Fastify plugin options
@@ -28,6 +28,11 @@ export interface PluginOptions {
      * Zod schemas for validating request input. Pass `true` to load from standard location (need to enable `@core/zod` plugin in schema.zmodel).
      */
     zodSchemas?: ModelZodSchema | boolean;
+
+    /**
+     * API format to use from `@zenstackhq/server/api`
+     */
+    api: HandleRequestFn;
 }
 
 /**
@@ -56,7 +61,7 @@ const pluginHandler: FastifyPluginCallback<PluginOptions> = (fastify, options, d
         }
         const query = request.query as Record<string, string>;
 
-        const response = await handleRequest({
+        const response = await options.api({
             method: request.method,
             path: (request.params as any)['*'],
             query,
