--- packages/server/src/openapi/utils.ts
@@ -1,19 +1,0 @@
-import { AUXILIARY_FIELDS } from '@zenstackhq/sdk';
-
-/**
- * Recursively strip auxiliary fields from the given data.
- */
-export function stripAuxFields(data: unknown) {
-    if (Array.isArray(data)) {
-        return data.forEach(stripAuxFields);
-    } else if (data && typeof data === 'object') {
-        for (const [key, value] of Object.entries(data)) {
-            if (AUXILIARY_FIELDS.includes(key)) {
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                delete (data as any)[key];
-            } else {
-                stripAuxFields(value);
-            }
-        }
-    }
-}
